Split up views tests
