+++ tests/integration/views/test_index_views.py
@@ -1,0 +1,9 @@
+# import pytest
+
+
+# pytestmark = pytest.mark.django_db
+
+
+def test_index(client):
+    response = client.get("/")
+    assert "Hello World" in response.rendered_content
